// src/app/generated/models/problem.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * RFC 9457 problem details (COMMON-API-002)
 */
export interface Problem {

  /**
   * A constant from common.ErrorCodes
   */
  code: string;
  correlationId: string;
  detail?: string;
  errors?: Array<{
'field'?: string;
'message'?: string;
}>;
  instance?: string;
  status: number;
  title: string;
  type: string;
}